:truck: Move JavaScript source to separate files
