+++ plugins/csp-demo-node-editor/CMakeLists.txt
@@ -14,6 +14,7 @@ endif()
 # build plugin -------------------------------------------------------------------------------------
 
 file(GLOB SOURCE_FILES src/*.cpp src/nodes/*.cpp)
+file(GLOB NODE_FILES src/nodes/*.js)
 
 # Resoucre files and header files are only added in order to make them available in your IDE.
 file(GLOB HEADER_FILES src/*.hpp)
@@ -21,6 +22,7 @@ file(GLOB HEADER_FILES src/*.hpp)
 add_library(csp-demo-node-editor SHARED
   ${SOURCE_FILES}
   ${HEADER_FILES}
+  ${NODE_FILES}
 )
 
 target_link_libraries(csp-demo-node-editor
@@ -31,12 +33,16 @@ target_link_libraries(csp-demo-node-editor
 # Add this Plugin to a "plugins" folder in your IDE.
 set_property(TARGET csp-demo-node-editor PROPERTY FOLDER "plugins")
 
+# We mark all node source files as "header" in order to make sure that no one tries to compile them.
+set_source_files_properties(${NODE_FILES} PROPERTIES HEADER_FILE_ONLY TRUE)
+
 # Make directory structure available in your IDE.
 source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}" FILES 
-  ${SOURCE_FILES} ${HEADER_FILES}
+  ${SOURCE_FILES} ${HEADER_FILES} ${NODE_FILES}
 )
 
 # install plugin -----------------------------------------------------------------------------------
 
-install(TARGETS   csp-demo-node-editor  DESTINATION "share/plugins")
+install(TARGETS csp-demo-node-editor  DESTINATION "share/plugins")
+install(FILES   ${NODE_FILES}         DESTINATION "share/resources/nodes/csp-demo-node-editor")
