+++ plugins/csp-demo-node-editor/src/nodes/TimeNode.cpp
@@ -10,6 +10,7 @@
 #include "../logger.hpp"
 
 #include "../../../../src/cs-core/TimeControl.hpp"
+#include "../../../../src/cs-utils/filesystem.hpp"
 #include "../../../../src/cs-utils/utils.hpp"
 
 namespace csp::demonodeeditor {
@@ -21,37 +22,8 @@ const std::string TimeNode::sName = "Time";
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 std::string TimeNode::sSource() {
-  return R"(
-    //js
-
-    // The TimeNode is pretty simple as it only has a single output socket. The component serves as
-    // a kind of factory. Whenever a new node is created, the builder() method is called.
-    class TimeComponent extends Rete.Component {
-
-      constructor() {
-        // This name must match the TimeNode::NAME defined above.
-        super("Time");
-
-        // This specifies the submenu from which this node can be created in the node editor.
-        this.category = "Inputs";
-      }
-
-      // Called whenever a new node of this type needs to be constructed.
-      builder(node) {
-
-        // This node has a single output. The first parameter is the name of this output and must be
-        // unique amongst all sockets. It is also used in the TimeNode::process() to write the
-        // output of this node. The second parameter is shown as name on the node. The last
-        // parameter references a socket type which has been registered with the node factory
-        // before. It is required that the class is called <NAME>Component.
-        let output = new Rete.Output('time', "Seconds", CosmoScout.socketTypes['Number Value']);
-        node.addOutput(output)
-
-        return node;
-      }
-    }
-    //!js
-  )";
+  return cs::utils::filesystem::loadToString(
+      "../share/resources/nodes/csp-demo-node-editor/TimeNode.js");
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
